refactor(client): providers, remove unneeded props and comments
