+++ packages/client/src/providers/SnackbarProvider/index.tsx
@@ -1,31 +1,20 @@
 import React, { createContext, FC, useRef, useState } from 'react';
 
-import { Snackbar } from '../elements';
-import { SnackbarSeverity } from '../elements/Snackbar/Snackbar.types';
-import useOnClickOutside from '../hooks/useOnClickOutside';
-
-export interface SnackbarContextValue {
-  showSnackbar: (props: ShowSnackbarProps) => void;
-}
+import { Snackbar } from '../../elements';
+import { SnackbarSeverity } from '../../elements/Snackbar/Snackbar.types';
+import useOnClickOutside from '../../hooks/useOnClickOutside';
+import getDisplayDuration from './getDisplayDuration';
+import {
+  ShowSnackbarProps,
+  SnackbarContextValue,
+  SnackbarProviderProps,
+} from './SnackbarProvider.types';
 
 export const SnackbarContext = createContext<SnackbarContextValue>({
   showSnackbar: () => {},
 });
 
-const getDisplayDuration = (message: string): number => {
-  const { min, max } = Math;
-  const msgLength = message.length;
-
-  return min(max(msgLength * 50, 2000), 7000);
-};
-
-export interface ShowSnackbarProps {
-  message: string;
-  severity?: SnackbarSeverity;
-  duration?: number;
-}
-
-const SnackbarProvider: FC = ({ children }) => {
+const SnackbarProvider = ({ children }: SnackbarProviderProps) => {
   const [isOpen, setOpen] = useState(false);
   const [messageInternal, setMessageInternal] = useState('');
   const [severityInternal, setSeverityInternal] = useState<SnackbarSeverity>('info');
