+++ packages/client/src/providers/ThemeTypeProvider/ThemeTypeProvider.types.d.ts
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react';
+
+import { ThemeType } from '../../theme/Theme';
+
+export interface ThemeTypeContextValue {
+  themeType: ThemeType;
+  setThemeType: (themeType: ThemeType) => void;
+}
+
+export interface ThemeProviderProps {
+  children: ReactNode;
+  themeType?: ThemeType;
+}
